# perf: Added benchmark test and limited expression compiling to once unless required