This is the earlier benchmark (LuaJIT) from GitHub Actions:
```
Raw:		Total ops = 10000		Time per op = 0.0005 ms
Basic Model:		Total ops = 10000		Time per op = 0.1157 ms
RBAC Model:		Total ops = 10000		Time per op = 0.2935 ms
RBAC Model Small:		Total ops = 1000		Time per op = 6.3911 ms
RBAC Model Medium:		Total ops = 100		Time per op = 63.5531 ms
RBAC Model Large:		Total ops = 10		Time per op = 670.6526 ms
RBAC Model with Resources:		Total ops = 1000		Time per op = 0.3454 ms
RBAC Model with Domains:		Total ops = 1000		Time per op = 0.4186 ms
ABAC Model:		Total ops = 1000		Time per op = 0.0311 ms
KeyMatch Model:		Total ops = 1000		Time per op = 0.4127 ms
RBAC Model with Deny:		Total ops = 1000		Time per op = 0.3503 ms
Priority Model:		Total ops = 1000		Time per op = 0.4927 ms
```
This is the current one:
```
Raw:		Total ops = 10000		Time per op = 0.0005 ms
Basic Model:		Total ops = 10000		Time per op = 0.0786 ms
RBAC Model:		Total ops = 10000		Time per op = 0.1393 ms
RBAC Model Small:		Total ops = 1000		Time per op = 1.1131 ms
RBAC Model Medium:		Total ops = 100		Time per op = 11.4998 ms
RBAC Model Large:		Total ops = 10		Time per op = 127.0854 ms
RBAC Model with Resources:		Total ops = 1000		Time per op = 0.2428 ms
RBAC Model with Domains:		Total ops = 1000		Time per op = 0.2113 ms
ABAC Model:		Total ops = 1000		Time per op = 0.0484 ms
KeyMatch Model:		Total ops = 1000		Time per op = 0.1511 ms
RBAC Model with Deny:		Total ops = 1000		Time per op = 0.3070 ms
Priority Model:		Total ops = 1000		Time per op = 0.2016 ms
```