@hsluoyz 2021-07-07
@rushitote plz also consider adding Lua result to: https://casbin.org/docs/en/benchmark , we want to add a multi-language tab like code snipet then add benchmarks for different languages.